# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reserva de turnos",
        "description": "Esta es la documentacion de nsdnasdkasnd.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/canchas/show/{id}": {
            "get": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Obtener detalles de una cancha",
                "description": "Retorna los detalles de una cancha específica.",
                "operationId": "f1f405f9af7d962c3c5347da5b2b9d94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la cancha a obtener.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la cancha obtenidos exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Nombre de la cancha",
                                            "nullable": true
                                        },
                                        "precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "10.5",
                                            "nullable": true
                                        },
                                        "techo": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "cant_jugadores": {
                                            "type": "integer",
                                            "example": 5,
                                            "nullable": true
                                        },
                                        "superficie": {
                                            "type": "string",
                                            "example": "Césped",
                                            "nullable": true
                                        },
                                        "id_categoria": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "activo": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cancha no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/canchas/categoria/{id_categoria}": {
            "get": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Buscar canchas por categoría",
                "description": "Retorna una lista de canchas que pertenecen a una categoría específica.",
                "operationId": "562d2ead5424321d0a06dde6c32797c6",
                "parameters": [
                    {
                        "name": "id_categoria",
                        "in": "path",
                        "description": "ID de la categoría para buscar las canchas.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de canchas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Nombre de la cancha",
                                                "nullable": true
                                            },
                                            "precio": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "10.5",
                                                "nullable": true
                                            },
                                            "techo": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "cant_jugadores": {
                                                "type": "integer",
                                                "example": 5,
                                                "nullable": true
                                            },
                                            "superficie": {
                                                "type": "string",
                                                "example": "Césped",
                                                "nullable": true
                                            },
                                            "id_categoria": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "activo": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron canchas para la categoría indicada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rest/categorias/show/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener detalles de una categoría",
                "description": "Retorna los detalles de una categoría específica.",
                "operationId": "1726d0a1f04922cf2e3842b2dad1f081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a obtener.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la categoría obtenidos exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Nombre de la categoría",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-05-18 12:34:56"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-05-18 12:34:56"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}